character literals: allow unicode escapes
also make the documentation for character literals more clear.
closes #2089

see #2097

+++ src/all_types.hpp
@@ -845,7 +845,7 @@ struct AstNodeStringLiteral {
 };
 
 struct AstNodeCharLiteral {
-    uint8_t value;
+    uint32_t value;
 };
 
 struct AstNodeFloatLiteral {
